requirements.txt
paho-mqtt
fritzconnection